examples/requirements.txt
fastapi == 0.101.0
uvicorn[standard] == 0.23.2
psycopg2-binary == 2.9.9
Jinja2 == 3.1.2